LayoutTests/http/tests/security/contentSecurityPolicy/eval-allowed-in-report-only-mode-and-sends-report.html

<!DOCTYPE html>
<html>
<head>
    <meta http-equiv="Content-Security-Policy-Report-Only" content="script-src 'self' 'unsafe-inline'; report-uri resources/save-report.php">
</head>
<body>
    <script>
        try {
            eval("alert('PASS: eval() allowed!')");
        } catch (e) {
            console.log('FAIL: eval() blocked!');
        }
    </script>
    <script src="resources/go-to-echo-report.js"></script>
</body>
</html>
